Use TextExtracts and PageImages extensions
TextExtracts provides clean plain text summaries of article context,
and PageImages selects useful images to be used for display, as can
be seen by their usage in Popups. This adds a hard dependency
upon both extensions.

The extension now creates "summary" type cards, which works best
for Wikipedia-type content. We'll probably want to add galleries
and photo cards for Commons and other projects.

An internal API call is made to communicate with TextExtracts
and PageImages.

$wgTwitterCardsHandle was added to specify the site's twitter handle
rather than falling back to $wgSitename.

This also fixes bug 62134 since we no longer use WikiPage (which was
never necessary to begin with).

Bug: 62134
Change-Id: I8e22a7a13aada2fb3b446923cec13b8496b9bc7a